[EI-125] [api] rename: api 폴더 경로 변경
