+++ api/src/community/api/dto/create-post.dto.ts
