+++ api/src/numerical-guidance/api/dto/create-indicator-board-meta-data.dto.ts
