+++ api/src/numerical-guidance/numerical-guidance.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { NumericalGuidanceController } from './infrastructure/api/numerical-guidance.controller';
+import { NumericalGuidanceController } from './api/numerical-guidance.controller';
 import { GetFluctuatingIndicatorQueryHandler } from './application/query/get-fluctuatingIndicator/get-fluctuatingIndicator.query.handler';
 import { FluctuatingIndicatorRedisAdapter } from './infrastructure/adapter/redis/fluctuatingIndicator.redis.adapter';
 import { FluctuatingIndicatorKrxAdapter } from './infrastructure/adapter/krx/fluctuatingIndicator.krx.adapter';
